Fixed setalias filter
Setalias filter updated from match exist to match all (match exist would end up returning the first module in any slot rather than the module in a specific slot).

If an empty slot is selected the syscfg API complains about a NULL input which is the result of the findhw function. That function doesn't return a bad status but if nothing is found the resourcehandle remains NULL causing downstream issues.